Got AtomicJS 1.1 relatively stable.  We can start testing it out now.
This update contains numerous fixes and completes the principal
development removing the setTimeout hacks from the binding updates.
Additionally, this update contains a new deferred view update feature
that significantly reduces the number of reflow and repaint operations
that occur in the browser.  In  the repeater performance demo, time
spent reflowing the layout for 1000 items was reduced from ~250ms to
~35 on my dev machine.