several bug fixes!
Reimagined the order and placement of framing and initialization of controls.  This new order unfortunately drops support for forwarding control properties to bound properties of child controls in composite controls, but there is a better pattern for doing that using the customBind feature.
Fixed an issue where computed properties defined on an observer fail to notify linked child observers of their initial value when the child observer is linked before the computed property is defined.  There my still be a similar issue related to bindings using nestedRootPaths that cover the location of the new computed property.  I will look into this next.